Move parse_bash_history_from_memory into shell backend
