+++ src/history/mod.rs
@@ -6,9 +6,6 @@ use std::vec;
 use crate::content::{self, StatefulSlidingWindow, apply_match_indices_to_lines};
 use crate::palette::Palette;
 use crate::shell;
-
-#[cfg(not(test))]
-use crate::shell::bash::symbols as bash_symbols;
 use flash::lexer::TokenKind;
 use itertools::Itertools;
 use ratatui::text::{Line, Span};
@@ -564,60 +561,6 @@ impl HistoryManager {
         merged
     }
 
-    #[cfg(test)]
-    pub fn parse_bash_history_from_memory() -> Vec<HistoryEntry> {
-        Vec::new()
-    }
-
-    #[cfg(not(test))]
-    pub fn parse_bash_history_from_memory() -> Vec<HistoryEntry> {
-        let mut res = Vec::with_capacity(4096);
-        unsafe {
-            let hist_array = bash_symbols::history_list();
-            if hist_array.is_null() {
-                log::warn!("History list is null");
-                return res;
-            }
-
-            let mut index = 0;
-            loop {
-                let entry_ptr = *hist_array.offset(index);
-                if entry_ptr.is_null() {
-                    break;
-                }
-
-                let hist_entry = &*entry_ptr;
-
-                // Check if line pointer is valid before dereferencing
-                if !hist_entry.line.is_null() {
-                    let command_cstr = std::ffi::CStr::from_ptr(hist_entry.line);
-                    let command_str = command_cstr.to_string_lossy().into_owned();
-
-                    // Parse timestamp if available
-                    let timestamp = if !hist_entry.timestamp.is_null() {
-                        let timestamp_cstr = std::ffi::CStr::from_ptr(hist_entry.timestamp);
-                        if let Ok(timestamp_str) = timestamp_cstr.to_str() {
-                            // If there are no timestamps in the history file,
-                            // Bash will use the current time for all entries, which can lead to many identical timestamps.
-                            HistoryManager::parse_timestamp(timestamp_str)
-                                .map(|s| TimestampNanos::from_seconds(s).raw_nanos())
-                        } else {
-                            None
-                        }
-                    } else {
-                        None
-                    };
-
-                    let entry = HistoryEntry::new(timestamp, index as usize, command_str);
-                    res.push(entry);
-                }
-
-                index += 1;
-            }
-        }
-        res
-    }
-
     fn parse_zsh_history(custom_path: Option<&str>) -> Vec<HistoryEntry> {
         let hist_path = match custom_path {
             Some(p) if !p.is_empty() => p.to_string(),
@@ -877,7 +820,7 @@ impl HistoryManager {
         self.last_word_insert_index = None;
     }
 
-    fn parse_timestamp(line: &str) -> Option<u64> {
+    pub(crate) fn parse_timestamp(line: &str) -> Option<u64> {
         // Bash writes `#<timestamp>` directly (digits immediately after `#` with no whitespace).
         let rest = line.strip_prefix('#')?;
         if !rest.is_empty() && rest.chars().all(|c| c.is_ascii_digit()) {
